+++ grafos/bin/.gitignore
@@ -1,0 +1,4 @@
+/grafo/
+/algoritmos/
+/principal/
+/algoritmo/
